+++ src/use/ValidationResponseHandler.ts
@@ -39,8 +39,6 @@ export default class ValidationResponseHandler implements ResponseHandler<boolea
                 return false;
             }
             case 422: {
-                console.log(response);
-                console.log(this.errorList);
                 return false;
             }
             case 201: {