+++ src/views/admin/Home.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script lang="ts">
-    import { store } from "../../store/store";
+    import { checkPrivilege } from "../../use/PermissionHelper";
     import { Role } from "../../entities/Role";
     import AdminAccountList from "../../components/AdminAccountList.vue";
 
@@ -18,12 +18,17 @@
             AdminAccountList,
         },
 
-        beforeRouteEnter(_to: any, _from: any, next: any) {
-            const myRole = store.state.myRole;
-            if (myRole != Role.ADMIN) {
-                return next("/redirect");
+        async beforeRouteEnter(_to: any, _from: any, next: any) {
+            const response = await checkPrivilege(Role.ADMIN);
+
+            if (response.allowed) {
+                return next();
             }
-            return next();
+            if (!response.authenticated) {
+                return next("/login");
+            }
+
+            return next("/redirect");
         },
     };
 </script>