+++ src/views/common/Login.vue
@@ -68,10 +68,10 @@
 </template>
 <script lang="ts">
     import Router from "@/router/";
-    import { store } from "../../store/store";
+    import { useStore, store } from "../../store/store";
     import { Role } from "../../entities/Role";
     import UserManagement from "@/api/UserManagement";
-    import { ref } from "vue";
+    import { ref, onMounted } from "vue";
     import LoginResponseHandler from "@/use/LoginResponseHandler";
 
     export default {
@@ -102,14 +102,13 @@
 
             async function login() {
                 const username = email.value;
-                const userManagement: UserManagement = new UserManagement();
-
-                const response = await userManagement.login({ username: username, password: password.value });
+                const response = await UserManagement.login({ username: username, password: password.value });
 
                 const loginSuccess = loginResponseHandler.handleReponse(response);
 
                 if (loginSuccess) {
-                    switch (store.state.myRole) {
+                    const store = useStore();
+                    switch (await store.getters.role) {
                         case Role.ADMIN: {
                             Router.push("/admin");
                             break;