+++ src/views/lecturer/Home.vue
@@ -9,7 +9,7 @@
 
 <script lang="ts">
     import CourseList from "../../components/LecturerCourseList.vue";
-    import { store } from "../../store/store";
+    import { checkPrivilege } from "../../use/PermissionHelper";
     import { Role } from "../../entities/Role";
 
     export default {
@@ -18,12 +18,17 @@
             CourseList,
         },
 
-        beforeRouteEnter(_to: any, _from: any, next: any) {
-            const myRole = store.state.myRole;
-            if (myRole != Role.LECTURER) {
-                return next("/redirect");
+        async beforeRouteEnter(_to: any, _from: any, next: any) {
+            const response = await checkPrivilege(Role.LECTURER);
+
+            if (response.allowed) {
+                return next();
             }
-            return next();
+            if (!response.authenticated) {
+                return next("/login");
+            }
+
+            return next("/redirect");
         },
     };
 </script>