+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Refactor
 - extract searchbar component for better adaptability (#131)
 - extract sections from account form in single components (#238)
+- bundle birthdate logic in birthdate picker component and remove the old birthdate object within the account (#241)
 
 ### Usability
 - rework login error validation (#232)