+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 - Add option for user to change the password (atm. in lecturer private profile, will be moved to a settings page) (#208)
 - Redesign admin account list to have a more modern look [#226](https://github.com/upb-uc4/ui-web/pull/226)
 - Allow re-authentication on page reload (#248)
+- Add settings page [#250](https://github.com/upb-uc4/ui-web/pull/250)
 
 ## Bug Fixes
 
 ## Refactor
 - extract searchbar component for better adaptability (#131)
 - extract sections from account form in single components (#238)
+- bundle birthdate logic in birthdate picker component and remove the old birthdate object within the account (#241)
 
 ### Usability
 - rework login error validation (#232)