+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Redesign admin account list to have a more modern look [#226](https://github.com/upb-uc4/ui-web/pull/226)
 - Allow re-authentication on page reload (#248)
 - Add settings page [#250](https://github.com/upb-uc4/ui-web/pull/250)
+- Add private profile page for admin (#252)
 
 ## Bug Fixes